blog_project/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
Markdown==3.5.2
Pillow==10.2.0
django-filter==23.5
django-cors-headers==4.3.1

